refactor: 도메인 추상화로 외부 라이브러리와의 의존성 역전
