+++ src/main/java/subway/domain/path/SectionEdge.java
@@ -1,0 +1,10 @@
+package subway.domain.path;
+
+import subway.domain.section.Section;
+
+public interface SectionEdge {
+
+    Section getSection();
+
+    Long getLineId();
+}
