+++ src/main/java/subway/jgrapht/JgraphtSectionEdge.java
@@ -1,21 +1,25 @@
-package subway.domain.section;
+package subway.jgrapht;
 
 import org.jgrapht.graph.DefaultWeightedEdge;
+import subway.domain.path.SectionEdge;
+import subway.domain.section.Section;
 
-public class SectionEdge extends DefaultWeightedEdge {
+public class JgraphtSectionEdge extends DefaultWeightedEdge implements SectionEdge {
 
     private final Section section;
     private final Long lineId;
 
-    public SectionEdge(Section section, Long lineId) {
+    public JgraphtSectionEdge(Section section, Long lineId) {
         this.section = section;
         this.lineId = lineId;
     }
 
+    @Override
     public Section getSection() {
         return section;
     }
 
+    @Override
     public Long getLineId() {
         return lineId;
     }
